change code organization: move _reload to new sub package _entity
