+++ src/taipy/core/_entity/_entity.py
@@ -9,7 +9,7 @@
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-from ..common._reload import _set_entity
+from src.taipy.core._entity._reload import _set_entity
 
 
 class _Entity:
