+++ src/taipy/core/data/json.py
@@ -18,10 +18,10 @@ from os.path import isfile
 from pydoc import locate
 from typing import Any, Dict, List, Optional, Set
 
+from src.taipy.core._entity._reload import _self_reload
 from taipy.config.common.scope import Scope
 
 from .._version._version_manager_factory import _VersionManagerFactory
-from ..common._reload import _self_reload
 from .abstract_file import _AbstractFileDataNode
 from .data_node import DataNode
 from .data_node_id import DataNodeId, Edit
