+++ src/taipy/core/task/task.py
@@ -12,6 +12,7 @@
 import uuid
 from typing import Any, Callable, Dict, Iterable, List, Optional, Set, Union
 
+from src.taipy.core._entity._reload import _reload, _self_reload, _self_setter
 from taipy.config.common._template_handler import _TemplateHandler as _tpl
 from taipy.config.common._validate_id import _validate_id
 from taipy.config.common.scope import Scope
@@ -19,7 +20,6 @@ from taipy.config.common.scope import Scope
 from .._entity._entity import _Entity
 from .._entity._properties import _Properties
 from .._version._version_manager_factory import _VersionManagerFactory
-from ..common._reload import _reload, _self_reload, _self_setter
 from ..common._warnings import _warn_deprecated
 from ..data.data_node import DataNode
 from .task_id import TaskId
